refactor: remove unused constants and components, streamline imports across the application
