+++ client/src/components/Footer/FooterBrand.jsx
@@ -1,6 +1,6 @@
 import React, { memo } from 'react';
-import { Facebook, Instagram, Twitter, Linkedin } from 'lucide-react';
-import { socialPlatforms } from '../../constants/SocialPlatforms';
+import { Facebook, Instagram, Twitter, Linkedin, Leaf } from 'lucide-react';
+import { socialPlatforms } from '../../constants';
 
 
 const iconMap = {
@@ -16,9 +16,7 @@ const FooterBrand = () => {
             <div className="flex items-center mb-6">
                 <div className="w-10 h-10 bg-gradient-to-br from-cambridge-blue-400 to-mindaro-500 rounded-lg flex items-center justify-center shadow-lg mr-3">
                     <div className="text-white">
-                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="w-6 h-6">
-                            <path d="M17,8C8,10 5.9,16.17 3.82,21.34L5.71,22L6.66,19.7C7.14,19.87 7.64,20 8,20C19,20 22,3 22,3C21,5 14,5.25 9,6.25C4,7.25 2,11.5 2,13.5C2,15.5 3.75,17.25 3.75,17.25C7,8 17,8 17,8Z" />
-                        </svg>
+                        <Leaf size={24} className="w-6 h-6" />
                     </div>
                 </div>
                 <div>
