--- client/src/constants/Marketplace/MarketplaceConstants.js
@@ -1,32 +1,0 @@
-export const categories = [
-  { value: "all", label: "All Categories" },
-  { value: "vegetables", label: "Vegetables" },
-  { value: "fruits", label: "Fruits" },
-  { value: "grains", label: "Grains & Rice" },
-  { value: "spices", label: "Spices" },
-  { value: "tea", label: "Tea & Coffee" },
-  { value: "coconut", label: "Coconut Products" },
-];
-
-export const locations = [
-  { value: "", label: "All Locations" },
-  { value: "kandy", label: "Kandy" },
-  { value: "colombo", label: "Colombo" },
-  { value: "galle", label: "Galle" },
-  { value: "jaffna", label: "Jaffna" },
-  { value: "nuwara-eliya", label: "Nuwara Eliya" },
-];
-
-export const sortOptions = [
-  { value: "newest", label: "Newest First" },
-  { value: "price-low", label: "Price: Low to High" },
-  { value: "price-high", label: "Price: High to Low" },
-  { value: "rating", label: "Highest Rated" },
-];
-
-export const certifications = [
-  { id: "organic", label: "Organic Certified" },
-  { id: "fair-trade", label: "Fair Trade" },
-  { id: "pesticide-free", label: "Pesticide-Free" },
-  { id: "export-quality", label: "Export Quality" },
-];
