--- client/src/constants/SocialPlatforms.js
@@ -1,20 +1,0 @@
-export const socialPlatforms = [
-  {
-    id: 1,
-    name: "Facebook",
-    icon: "📘",
-    url: "https://facebook.com",
-  },
-  {
-    id: 2,
-    name: "Twitter",
-    icon: "🐦",
-    url: "https://twitter.com",
-  },
-  {
-    id: 3,
-    name: "Instagram",
-    icon: "📷",
-    url: "https://instagram.com",
-  },
-];
