--- client/src/constants/legalLinks.js
@@ -1,17 +1,0 @@
-export const legalLinks = [
-  {
-    id: 1,
-    label: "Terms of Service",
-    path: "/terms",
-  },
-  {
-    id: 2,
-    label: "Privacy Policy",
-    path: "/privacy",
-  },
-  {
-    id: 3,
-    label: "Cookie Policy",
-    path: "/cookies",
-  },
-];
