--- client/src/constants/navItems.js
@@ -1,17 +1,0 @@
-export const navItems = [
-  {
-    id: 1,
-    label: "Home",
-    path: "/",
-  },
-  {
-    id: 2,
-    label: "Marketplace",
-    path: "/marketplace",
-  },
-  {
-    id: 3,
-    label: "About",
-    path: "/about",
-  },
-];
