--- client/src/constants/steps.js
@@ -1,60 +1,0 @@
-export const steps = [
-  {
-    id: 0,
-    number: "01",
-    title: "Create Your Profile",
-    description:
-      "Get started with CropMate by creating your personalized profile:",
-    features: [
-      "Register as a Farmer, Vendor, or Transporter",
-      "Add your location and business details",
-      "Upload verification documents",
-      "Set your preferences and specialties",
-    ],
-    image: "/src/assets/images/create-profile.jpg",
-    color: "cambridge-blue",
-    iconBg: "bg-cambridge-blue-500",
-    decorColor:
-      "bg-gradient-to-r from-cambridge-blue-400/10 to-cambridge-blue-500/5",
-    buttonText: "Sign Up",
-    buttonLink: "/signup",
-  },
-  {
-    id: 1,
-    number: "02",
-    title: "Connect & Trade",
-    description: "Connect with partners across the agricultural ecosystem:",
-    features: [
-      "Browse listings by location and category",
-      "Message potential partners securely",
-      "Compare prices and quality ratings",
-      "Negotiate and agree on terms",
-    ],
-    image: "/src/assets/images/connect-trade.jpg",
-    color: "golden-brown",
-    iconBg: "bg-golden-brown-500",
-    decorColor:
-      "bg-gradient-to-r from-golden-brown-400/10 to-golden-brown-500/5",
-    buttonText: "Browse Crops",
-    buttonLink: "/marketplace",
-  },
-  {
-    id: 2,
-    number: "03",
-    title: "Grow Together",
-    description: "Build sustainable partnerships that grow your business:",
-    features: [
-      "Leave reviews for completed transactions",
-      "Build reputation and earn badges",
-      "Access data insights and trends",
-      "Participate in the CropMate community",
-    ],
-    image: "/src/assets/images/grow-together.jpg",
-    color: "cal-poly-green",
-    iconBg: "bg-cal-poly-green-300",
-    decorColor:
-      "bg-gradient-to-r from-cal-poly-green-400/10 to-cal-poly-green-500/5",
-    buttonText: "Learn More",
-    buttonLink: "/about",
-  },
-];
